kira-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
redis>=5.2.0
chromadb>=0.5.0
google-genai>=1.0.0
PyJWT>=2.10.0
prometheus-client>=0.21.0
httpx>=0.27.0
openai>=1.50.0
